# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Journalists Quotes Service",
    "description": "Three workflows: (WF1) email inbound HARO ingestion -> silver+gold; (WF2) /orgs/opportunities/next merges silver email opps with live Featured fetch + RAG scoring; (WF3) /orgs/opportunities/{id}/reply dispatches via Featured submitAnswer or email-gateway send.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3050"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "OpportunityNextResponse": {
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "match"
                ]
              },
              "opportunityId": {
                "type": "string",
                "format": "uuid"
              },
              "provider": {
                "type": "string"
              },
              "ingestionChannel": {
                "type": "string"
              },
              "featuredQuestionId": {
                "type": "integer",
                "nullable": true
              },
              "mediaOutlet": {
                "type": "string",
                "nullable": true
              },
              "journalistName": {
                "type": "string",
                "nullable": true
              },
              "opportunityText": {
                "type": "string"
              },
              "deadline": {
                "type": "string",
                "nullable": true
              },
              "pitchUrl": {
                "type": "string",
                "nullable": true
              },
              "pitchEmail": {
                "type": "string",
                "nullable": true
              },
              "score": {
                "type": "number"
              },
              "whyRelevant": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "status",
              "opportunityId",
              "provider",
              "ingestionChannel",
              "featuredQuestionId",
              "mediaOutlet",
              "journalistName",
              "opportunityText",
              "deadline",
              "pitchUrl",
              "pitchEmail",
              "score",
              "whyRelevant"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "no_match"
                ]
              }
            },
            "required": [
              "status"
            ]
          }
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "OpportunityNextRequest": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "brandId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "campaignId",
          "brandId"
        ]
      },
      "OpportunityReplyResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "submitted",
              "already_submitted",
              "rate_limited",
              "error"
            ]
          },
          "pitchId": {
            "type": "string",
            "format": "uuid"
          },
          "deliveryMethod": {
            "type": "string",
            "enum": [
              "featured_api",
              "email_reply"
            ]
          },
          "outboundMessageId": {
            "type": "string",
            "nullable": true
          },
          "featuredQuestionId": {
            "type": "integer",
            "nullable": true
          },
          "retryAfter": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "OpportunityReplyRequest": {
        "type": "object",
        "properties": {
          "pitchContent": {
            "type": "string",
            "minLength": 1
          },
          "brandId": {
            "type": "string",
            "format": "uuid"
          },
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "subject": {
            "type": "string"
          }
        },
        "required": [
          "pitchContent",
          "brandId",
          "campaignId"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/orgs/opportunities/next": {
      "post": {
        "summary": "Pick the next best journalist opportunity for the (campaign, brand). Merges silver email-sourced rows with a live fetch of Featured.com opportunities, scores via RAG, returns top above SCORE_THRESHOLD.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpportunityNextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Top opportunity or no_match",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityNextResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream service unavailable (key-service, Featured)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/opportunities/{id}/reply": {
      "post": {
        "summary": "Submit a pitch reply for the given opportunity. Dispatches to Featured submitAnswer (provider=featured) or email-gateway-service /orgs/send (other providers, e.g. haro).",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpportunityReplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pitch dispatched (or idempotent already-submitted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityReplyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or unsupported opportunity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Opportunity not found for this org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}